[proof] More uniformity on iterators, proof types.
The type invariants here are:

   leaf counts are always usize.  This is because counts can validly be
   zero.  Functions which do not function for zero leaf counts should
   assert so.

   leaf indices are always u64.  The index always starts from zero.
   If there is a valid need to represent an invalid index, use
   Option<u64> or use a count of leaves instead.

   Version is a slightly more abstract type; it could reasonably
   be changed to a tuple or some other struct in the future.  For
   now, try to avoid arithmetic on a version by converting it into
   an index (u64) first.

+++ storage/accumulator/src/tests/write_test.rs
@@ -22,8 +22,7 @@ fn test_append_one() {
     let mut leaves = Vec::new();
     for _ in 0..100 {
         let hash = HashValue::random();
-        let (root_hash, writes) =
-            TestAccumulator::append(&store, leaves.len() as u64, &[hash]).unwrap();
+        let (root_hash, writes) = TestAccumulator::append(&store, leaves.len(), &[hash]).unwrap();
         store.put_many(&writes);
 
         leaves.push(hash);
@@ -47,7 +46,7 @@ proptest! {
                 TestAccumulator::append(&store, num_leaves, &hashes).unwrap();
             store.put_many(&writes);
 
-            num_leaves += hashes.len() as u64;
+            num_leaves += hashes.len();
             leaves.extend(hashes.iter());
             let expected_root_hash = store.verify(&leaves).unwrap();
             assert_eq!(root_hash, expected_root_hash)
@@ -62,7 +61,7 @@ proptest! {
         store.put_many(&writes);
 
         let (root_hash2, writes2) =
-            TestAccumulator::append(&store, leaves.len() as u64, &[]).unwrap();
+            TestAccumulator::append(&store, leaves.len(), &[]).unwrap();
 
         assert_eq!(root_hash, root_hash2);
         assert!(writes2.is_empty());
